+++ crates/soundlog/src/lib.rs
@@ -90,4 +90,4 @@ pub mod vgm;
 
 pub use binutil::ParseError;
 pub use vgm::command::*;
-pub use vgm::{VgmBuilder, VgmDocument, VgmExtraHeader, VgmHeader};
+pub use vgm::{VgmBuilder, VgmDocument, VgmExtraHeader, VgmHeader, VgmStream};